test(e2e): share native dialog capture helper
